implemented all the required changes
