+++ src/App.css
@@ -1,43 +1,57 @@
-#root {
-  display: flex;
-  flex-direction: column;
-  gap: 2em;
+body {
+  font-family: Arial, sans-serif;
+  margin: 0;
+  padding: 0;
 }
 
 header {
+  background-color: #333;
+  color: white;
+  text-align: center;
+  padding: 1rem;
+}
+
+nav ul {
+  list-style: none;
   display: flex;
-  justify-content: space-between;
-  margin: 2%;
 }
 
-header ul {
+nav li {
+  margin-right: 1rem;
+}
+
+nav a {
+  text-decoration: none;
+  color: #333;
+}
+
+.book {
+  border: 1px solid #ccc;
+  padding: 1rem;
+  margin: 1rem;
+}
+
+.form {
   display: flex;
-  gap: 2em;
+  flex-direction: column;
+  margin-top: 1rem;
 }
 
-.books {
-  margin: 2%;
+.form input {
+  margin-bottom: 0.5rem;
+  padding: 0.5rem;
 }
 
-.books li {
+.form button {
+  background-color: #333;
+  color: white;
+  padding: 0.5rem;
+  border: none;
+  cursor: pointer;
+}
+
+.books-container {
   display: flex;
-  justify-content: space-between;
-  margin-top: 1em;
-}
-
-.books div {
-  display: flex;
-  justify-content: space-between;
-  min-width: 50%;
-}
-
-form {
-  display: flex;
-  justify-content: center;
-  gap: 0.5em;
-  margin-left: 2%;
-}
-
-span {
-  font-weight: 600;
+  flex-direction: column;
+  align-items: center;
 }
