+++ src/App.js
@@ -1,6 +1,7 @@
+import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import Header from './components/Navbar';
-import Books from './components/Book';
+import Navbar from './components/Navbar';
+import Books from './components/Books';
 import Categories from './components/Categories';
 import Form from './components/Form';
 import './App.css';
@@ -8,7 +9,7 @@ import './App.css';
 function App() {
   return (
     <BrowserRouter>
-      <Header />
+      <Navbar />
       <Routes>
         <Route path="/" element={<Books />} />
         <Route path="/categories" element={<Categories />} />
