+++ src/components/Book.js
@@ -1,30 +1,17 @@
-import React, { useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
-import Book from './Book';
-import Form from './Form';
+import React from 'react';
 
-const Books = () => {
-  const [books, setBooks] = useState([
-    { id: uuidv4(), title: 'The Great Gatsby', author: 'Scott Fitzgerald' },
-    { id: uuidv4(), title: 'Moby Dick', author: 'Herman Melville' },
-    { id: uuidv4(), title: 'War and Peace', author: 'Leo Tolstoy' },
-    { id: uuidv4(), title: 'The Brothers Karamazov', author: 'Fyodor Dostoyevsky' },
-  ]);
-
-  const addBook = (newBook) => {
-    setBooks([...books, { ...newBook, id: uuidv4() }]);
+const Book = ({ id, title, author, onRemove }) => {
+  const handleRemove = () => {
+    onRemove(id);
   };
 
   return (
-    <>
-      <ul className="books">
-        {books.map((book) => (
-          <Book key={book.id} title={book.title} author={book.author} />
-        ))}
-      </ul>
-      <Form addBook={addBook} />
-    </>
+    <li className="book">
+      <h3>Title: {title}</h3>
+      <p>Author: {author}</p>
+      <button onClick={handleRemove}>Remove</button>
+    </li>
   );
 };
 
-export default Books;
+export default Book;
